requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckiptagger==0.2.1
click==8.1.7
colorama==0.4.6
datasets==2.21.0
dill==0.3.8
evaluate==0.4.3
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.4.0
gdown==5.2.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.11.0
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
nltk==3.9.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rouge==1.0.1
rsa==4.9
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.1.0+cu121
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
